+++ app/build.gradle
@@ -55,7 +55,6 @@ dependencies {
     androidTestImplementation 'androidx.test.ext:junit:1.1.5'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
     implementation 'com.google.code.gson:gson:2.10.1'
-    //Retrofit
     implementation "com.squareup.retrofit2:retrofit:2.9.0"
     implementation "com.squareup.retrofit2:converter-gson:2.6.0"
     implementation "com.squareup.retrofit2:adapter-rxjava2:2.9.0"
@@ -65,5 +64,6 @@ dependencies {
     implementation "com.squareup.picasso:picasso:2.71828"
     implementation 'androidx.paging:paging-common-ktx:3.2.1'
     implementation 'androidx.paging:paging-runtime-ktx:3.2.1'
+    implementation 'org.koin:koin-android:2.0.1'
 
 }
